+++ kdevplatform/language/util/kdevhash.h
@@ -15,7 +15,7 @@
  * @code
  * class MyType {
  *     ...
- *     uint hash() const
+ *     size_t hash() const
  *     {
  *         KDevHash hash;
  *         return hash << m_1 << m_2 << ...;
@@ -30,14 +30,14 @@ public:
         DEFAULT_SEED = 2166136261u
     };
 
-    explicit KDevHash(uint hash = DEFAULT_SEED)
+    explicit KDevHash(size_t hash = DEFAULT_SEED)
         : m_hash(hash)
     {}
 
     KDevHash(const KDevHash&) = delete;
     KDevHash& operator=(const KDevHash&) = delete;
 
-    inline operator uint() const
+    inline operator size_t() const
     {
         return m_hash;
     }
@@ -49,14 +49,14 @@ public:
         return *this;
     }
 
-    static inline uint hash_combine(uint seed, uint hash)
+    static inline size_t hash_combine(size_t seed, size_t hash)
     {
         // this is copied from boost::hash_combine
         return seed ^ (hash + 0x9e3779b9 + (seed << 6) + (seed >> 2));
     }
 
 private:
-    uint m_hash;
+    size_t m_hash;
 };
 
 #endif //KDEVPLATFORM_KDEVHASH_H
